// js/stat-evaluation-player/src/generated/BallHalfEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * A change in which half of the field the ball occupies.
 */
export type BallHalfEvent = { time: number, frame: number, end_time: number, end_frame: number, active: boolean, duration: number, field_half: string, };